move log initialisation
